+++ controller/editorcontroller.php
@@ -614,13 +614,14 @@ class EditorController extends Controller {
 	 *
 	 * @param array $referenceData - reference data
 	 * @param string $path - file path
+	 * @param string $link - file link
 	 *
 	 * @return array
 	 *
 	 * @NoAdminRequired
 	 * @PublicPage
 	 */
-	public function reference($referenceData, $path = null) {
+	public function reference($referenceData, $path = null, $link = null) {
 		$this->logger->debug("reference: " . json_encode($referenceData) . " $path", ["app" => $this->appName]);
 
 		if (!$this->config->isUserAllowedToUse()) {
@@ -652,6 +653,15 @@ class EditorController extends Controller {
 				&& $node->isReadable()
 			) {
 				$file = $node;
+			}
+		}
+
+		if ($file === null
+			&& !empty($link)
+		) {
+			$fileId = $this->getFileIdByLink($link);
+			if (!empty($fileId)) {
+				list($file, $error, $share) = $this->getFile($userId, $fileId);
 			}
 		}
 
@@ -1570,6 +1580,38 @@ class EditorController extends Controller {
 			$userId = end($userIdExp);
 		}
 		return $userId;
+	}
+
+	/**
+	 * Get File id from by link
+	 *
+	 * @param string $link - link to the file
+	 *
+	 * @return string|null
+	 */
+	private function getFileIdByLink(string $link) {
+		$path = parse_url($link, PHP_URL_PATH);
+		$encodedPath = array_map("urlencode", explode("/", $path));
+		$link = str_replace($path, implode("/", $encodedPath), $link);
+		if (filter_var($link, FILTER_VALIDATE_URL) === false) {
+			return null;
+		}
+
+		if (!empty($this->config->getStorageUrl())) {
+			$storageUrl = $this->config->getStorageUrl();
+		} else {
+			$storageUrl = $this->urlGenerator->getAbsoluteURL("/");
+		}
+
+		if (parse_url($link, PHP_URL_HOST) !== parse_url($storageUrl, PHP_URL_HOST)) {
+			return null;
+		}
+
+		if (preg_match('/\/onlyoffice\/(\d+)/', $link, $matches)) {
+			return $matches[1];
+		}
+
+		return null;
 	}
 
 	/**